[PGO][HIP] Split HSA device drain into its own file
Move the Linux-only supplemental HSA-introspection drain out of
InstrProfilingPlatformROCm.cpp into InstrProfilingPlatformROCmHSA.cpp,
with a private InstrProfilingPlatformROCmInternal.h declaring the shared
__prof_rocm interface (HIP helpers, processDeviceOffloadPrf, UniqueFree,
and the drained-bounds dedup hooks).

The new file is guarded by __linux__ && !_WIN32 and compiles to an empty
TU elsewhere; there is no Windows HSA path (Windows uses only the
host-shadow HIP drain). No behavior change; wired into CMake and the
Bazel overlay.
